dom/credentialmanagement/identity/tests/mochitest/test_accounts_redirect.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Server Redirect On Accounts Endpoint</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    setupTest("accounts_redirect").then(
      function () {
        return navigator.credentials.get({
          identity: {
            providers: [{
              configURL: "https://example.net/tests/dom/credentialmanagement/identity/tests/mochitest/server_manifest.sjs",
              clientId: "mochitest",
              nonce: "nonce"
            }]
          }
        });
      }
    ).then((cred) => {
      ok(false, "incorrectly got a credential");
    }).catch((err) => {
      ok(true, "correctly got an error");
    }).finally(() => {
      SimpleTest.finish();
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">This test verifies that we do not get a credential when the accounts endpoint redirects to another (entirely functional) accounts endpoint.</div>
<pre id="test"></pre>
</body>
</html>
